upnpav: Move Protocol in it's own file
