+++ upnpav/src/GetProtocolInfoResponse.hpp
@@ -4,6 +4,7 @@
 
 #pragma once
 
+#include "Protocol.hpp"
 #include "ServiceControlPointDefinition.hpp"
 #include "blabbyupnpav_export.h"
 #include <QString>
@@ -14,22 +15,7 @@ namespace UPnPAV
 {
 
 /**
- * The Protocol describes one specific protocol supported by the device. An asterix(*) is a wildcared and matches with
- * everything during string comparsions to determine the supported protocol in the control point. struct Protocol
- */
-struct BLABBYUPNPAV_EXPORT Protocol
-{
-    QString protocol;
-    QString network;
-    QString contentFormat;
-    QString additionalInfo;
-
-    BLABBYUPNPAV_EXPORT friend bool operator==(const Protocol &lhs, const Protocol &rhs) noexcept;
-    BLABBYUPNPAV_EXPORT friend bool operator!=(const Protocol &lhs, const Protocol &rhs) noexcept;
-};
-
-/**
- * This struct is the result of ConnectionManager::GetProtocolInfo call on a UPnPAV media device. The result contains
+ * This is the result of ConnectionManager::GetProtocolInfo call on a UPnPAV media device. The result contains
  * the supported source and sink protocols of the device.
  */
 class BLABBYUPNPAV_EXPORT GetProtocolInfoResponse
